% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.fairpolicy}
\alias{print.fairpolicy}
\title{Print a fitted fairpolicy object.}
\usage{
\method{print}{fairpolicy}(x, ...)
}
\arguments{
\item{x}{A fitted \code{fairpolicy} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fitted fairpolicy object.
}
