% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.fairpolicy_frontier}
\alias{print.fairpolicy_frontier}
\title{Print a fairpolicy Pareto frontier object.}
\usage{
\method{print}{fairpolicy_frontier}(x, ...)
}
\arguments{
\item{x}{A \code{fairpolicy_frontier} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fairpolicy Pareto frontier object.
}
