% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.fairtargeting}
\alias{print.fairtargeting}
\title{Print a fitted fairtargeting object.}
\usage{
\method{print}{fairtargeting}(x, ...)
}
\arguments{
\item{x}{A fitted \code{fairtargeting} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fitted fairtargeting object.
}
