% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.fairtargeting_frontier}
\alias{print.fairtargeting_frontier}
\title{Print a fairtargeting Pareto frontier object.}
\usage{
\method{print}{fairtargeting_frontier}(x, ...)
}
\arguments{
\item{x}{A \code{fairtargeting_frontier} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fairtargeting Pareto frontier object.
}
