% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.summary.fairpolicy}
\alias{print.summary.fairpolicy}
\title{Print a fairpolicy summary object.}
\usage{
\method{print}{summary.fairpolicy}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fairpolicy} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fairpolicy summary object.
}
