% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.summary.fairtargeting}
\alias{print.summary.fairtargeting}
\title{Print a fairtargeting summary object.}
\usage{
\method{print}{summary.fairtargeting}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fairtargeting} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fairtargeting summary object.
}
