% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.fairpolicy}
\alias{summary.fairpolicy}
\title{Summarize a fitted fairpolicy object.}
\usage{
\method{summary}{fairpolicy}(object, ...)
}
\arguments{
\item{object}{A fitted \code{fairpolicy} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{summary.fairpolicy}.
}
\description{
Summarize a fitted fairpolicy object.
}
