% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.fairtargeting}
\alias{summary.fairtargeting}
\title{Summarize a fitted fairtargeting object.}
\usage{
\method{summary}{fairtargeting}(object, ...)
}
\arguments{
\item{object}{A fitted \code{fairtargeting} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{summary.fairtargeting}.
}
\description{
Summarize a fitted fairtargeting object.
}
